+++ udata_front/frontend/helpers.py
@@ -17,6 +17,14 @@ from . import front
 
 from udata.core.dataset.apiv2 import dataset_fields
 from udata.core.dataset.models import Dataset
+from udata.core.dataservices.models import Dataservice
+from udata.core.organization.models import Organization
+from udata.core.organization.constants import (
+    ASSOCIATION,
+    COMPANY,
+    LOCAL_AUTHORITY,
+    PUBLIC_SERVICE,
+)
 from udata.models import db
 from udata.i18n import get_locale, format_date, format_timedelta, _, pgettext
 from udata.search.result import SearchResult
@@ -212,6 +220,20 @@ def owner_name_acronym(obj):
 
 
 @front.app_template_global()
+def organization_type(org: Organization):
+    if org.local_authority:
+        return LOCAL_AUTHORITY
+    elif org.public_service:
+        return PUBLIC_SERVICE
+    elif org.association:
+        return ASSOCIATION
+    elif org.company:
+        return COMPANY
+    else:
+        return ""
+
+
+@front.app_template_global()
 def external_source(dataset):
     return dataset.harvest.remote_url if dataset.harvest else None
 
@@ -398,6 +420,11 @@ def to_api_format(data):
 
 def to_dataset_api_format(dataset):
     return marshal(dataset, dataset_fields)
+
+
+@front.app_template_filter()
+def to_dataservice_api_format(dataservice):
+    return marshal(dataservice, Dataservice.__read_fields__)
 
 
 @front.app_template_filter()