+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## Current (in progress)
 
+- add dataservices to dataset page [#476](https://github.com/datagouv/udata-front/pull/476)
 - show contact point in dataset and dataservice page [#479](https://github.com/datagouv/udata-front/pull/479)
 - add email, member since and last login at to org members table [#480](https://github.com/datagouv/udata-front/pull/480)
+- always show authorization_request_url if present (even on open API) [#481](https://github.com/datagouv/udata-front/pull/481)
+- Show organization type [#472](https://github.com/datagouv/udata-front/pull/472)
 
 ## 5.1.2 (2024-08-01)